Collatz orbit calculator unit in Logisim
This contains an orbit calculator unit working with 16-bit registers. It
takes a number N as input and calculates a single Collatz orbit number
each clock cycle. When the number in internal register reaches 1, it
starts ignoring the clock cycles and does nothing. By that time, the
output pins will contain a length of an orbit and an orbit's "record
number", i.e. the peak value the orbit has reached before collapsing.

It also contains an array of 32 such units calculating the orbits
simultaneously (with a caveat that it's very cumbersome to use: you have
to set up the 32 inputs manually and read the results manually).